Update for initial batch size
