+++ Makefile
@@ -1,0 +1,5 @@
+coverage:
+	pytest --cov=scripts tests/
+
+test:
+	pytest
