+++ tests/test_script.py
@@ -1,0 +1,39 @@
+# import pytest
+
+import webui_daam.daam
+import scripts
+
+
+def test_show():
+    script = daam.Script()
+    assert script.show(is_img2img=False) == scripts.AlwaysVisible
+    assert script.show(is_img2img=True) == scripts.AlwaysVisible
+
+
+def test_process():
+    script = daam.Script()
+
+    # attention_texts: str,
+    # enabled: bool,
+    # show_images: bool,
+    # save_images: bool,
+    # show_caption: bool,
+    # use_grid: bool,
+    # grid_layout: str,
+    # alpha: float,
+    # heatmap_image_scale: float,
+    # trace_each_layers: bool,
+    # layers_as_row: bool,
+    script.process(
+        p=processing,
+        attention_text="word",
+        show_images=True,
+        save_images=False,
+        show_caption=False,
+        use_grid=False,
+        grid_layout="Auto",
+        alpha=0.5,
+        heatmap_image_scale=1.0,
+        trace_each_layers=False,
+        layers_as_row=False,
+    )
